Rename selector.py to selectors.py
